feat: start migrate on ts

+++ src/resources/boards/board.service.ts
@@ -1,15 +1,21 @@
-const boardsRepo = require('./board.memory.repository');
-const tasksRepo = require('../tasks/task.memory.repository');
-const Board = require('./board.model');
+import express from 'express';
+import * as boardsRepo from './board.memory.repository';
+import * as tasksRepo from '../tasks/task.memory.repository';
+import { Board } from './board.model';
 
-const getBoards = async (req, res) => {
+const getBoards = async (res: express.Response) => {
   const boards = await boardsRepo.getAll();
 
   res.status(200).json(boards.map(Board.toResponse));
 };
 
-const getBoard = async (req, res) => {
+const getBoard = async (req: express.Request, res: express.Response) => {
   const { id } = req.params;
+
+  if (!id) {
+    throw new Error('Error data');
+  }
+
   const board = await boardsRepo.getBoard(id);
 
   if (board) res.status(200).json(Board.toResponse(board));
@@ -17,7 +23,7 @@ const getBoard = async (req, res) => {
   if (!board) res.status(404).json({ message: 'Board not found' });
 };
 
-const createBoard = async (req, res) => {
+const createBoard = async (req: express.Request, res: express.Response) => {
   const newBoardInfo = req.body;
 
   const newBoard = await boardsRepo.createBoard(newBoardInfo);
@@ -25,17 +31,25 @@ const createBoard = async (req, res) => {
   if (newBoard) res.status(201).json(Board.toResponse(newBoard));
 };
 
-const updateBoard = async (req, res) => {
+const updateBoard = async (req: express.Request, res: express.Response) => {
   const { id } = req.params;
   const newInfo = req.body;
+
+  if (!id) {
+    throw new Error('Error data');
+  }
 
   const updatedBoard = await boardsRepo.updateBoard(id, newInfo);
 
   if (updatedBoard) res.status(200).json(Board.toResponse(updatedBoard));
 };
 
-const deleteBoard = async (req, res) => {
+const deleteBoard = async (req: express.Request, res: express.Response) => {
   const { id } = req.params;
+
+  if (!id) {
+    throw new Error('Error data');
+  }
 
   await tasksRepo.deleteTaskIfBoardDelete(id);
   const deletedBoard = await boardsRepo.deleteBoard(id);
@@ -46,4 +60,4 @@ const deleteBoard = async (req, res) => {
   if (!deletedBoard) res.status(404).json({ message: 'Board not found' });
 };
 
-module.exports = { getBoards, getBoard, createBoard, updateBoard, deleteBoard };
+export { getBoards, getBoard, createBoard, updateBoard, deleteBoard };
