+++ src/resources/tasks/task.service.ts
@@ -1,26 +1,31 @@
-const tasksRepo = require('./task.memory.repository');
-const Task = require('./task.model');
+import express from 'express';
+import * as tasksRepo from './task.memory.repository';
+import { Task } from './task.model';
 
-const getTasks = async (req, res) => {
+const getTasks = async (res: express.Response) => {
   const tasks = await tasksRepo.getAll();
 
   res.status(200).json(tasks.map(Task.toResponse));
 };
 
-const getTask = async (req, res) => {
+const getTask = async (req: express.Request, res: express.Response) => {
   const { id, boardId } = req.params;
+
+  if (!id || !boardId) {
+    throw new Error('Error data');
+  }
 
   const task = await tasksRepo.getTask(id, boardId);
 
   if (task) res.status(200).json(Task.toResponse(task));
 };
 
-const createTask = async (req, res) => {
+const createTask = async (req: express.Request, res: express.Response) => {
   const newTaskInfo = req.body;
 
   const boardId = newTaskInfo.boardId
     ? newTaskInfo.boardId
-    : req.params.boardId;
+    : req.params['boardId'];
 
   const newTask = await tasksRepo.createTask(boardId, newTaskInfo);
 
@@ -29,9 +34,13 @@ const createTask = async (req, res) => {
   if (newTask) res.status(201).json(Task.toResponse(newTask));
 };
 
-const updateTask = async (req, res) => {
+const updateTask = async (req: express.Request, res: express.Response) => {
   const { id, boardId } = req.params;
   const updateInfo = req.body;
+
+  if (!id || !boardId) {
+    throw new Error('Error data');
+  }
 
   const updatedTask = await tasksRepo.updateTask(id, boardId, updateInfo);
 
@@ -40,8 +49,12 @@ const updateTask = async (req, res) => {
   if (updatedTask) res.status(200).json(Task.toResponse(updatedTask));
 };
 
-const deleteTask = async (req, res) => {
+const deleteTask = async (req: express.Request, res: express.Response) => {
   const { id, boardId } = req.params;
+
+  if (!id || !boardId) {
+    throw new Error('Error data');
+  }
 
   const deletedTask = await tasksRepo.deleteTask(id, boardId);
 
@@ -51,10 +64,4 @@ const deleteTask = async (req, res) => {
   if (!deletedTask) res.status(404).json({ message: 'Task not found' });
 };
 
-module.exports = {
-  getTasks,
-  getTask,
-  createTask,
-  updateTask,
-  deleteTask,
-};
+export { getTasks, getTask, createTask, updateTask, deleteTask };
